requirements.txt
# ── Scraping ──────────────────────────────────────────────────────────────────
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
urllib3==2.2.3

# ── Telegram bot ──────────────────────────────────────────────────────────────
python-telegram-bot==21.9

# ── Scheduling ────────────────────────────────────────────────────────────────
schedule==1.2.2

# ── Config / utils ────────────────────────────────────────────────────────────
python-dotenv==1.0.1

